Move git app key coverage into session integration test.
Delete the standalone provider_local_keys_test.go and cover the ciCd
git source.path matching case via prepareProviderLocalSession instead.
